Struct agilulf_protocol::AsyncReadBuffer

pub struct AsyncReadBuffer<T: AsyncRead + Unpin> { /* fields omitted */ }

A Read buffer to provide read_exact and read_line methods.

Amazingly there isn't an AsyncBuffer for latest futures! The implementation of this struct is similar with BufferReader in std. The implementation of read_exact and read_line learns from BufferRead trait in std.

Methods

impl<T: AsyncRead + Unpin> AsyncReadBuffer<T>

pub fn new(stream: T) -> AsyncReadBuffer<T>

pub async fn fill_buf<'_>(&'_ mut self) -> Result<&'_ [u8]>

pub fn consume(&mut self, amt: usize)

pub async fn read_line<'_>(&'_ mut self) -> Result<Vec<u8>>

Note: The return value of this method contains "\r\n".

This method will find "\n" and check whether the former one byte is "\r". If it is, it will be consumed and return.

pub async fn read_exact<'_>(&'_ mut self, size: usize) -> Result<Vec<u8>>

impl<T: AsyncRead + Unpin + 'static> AsyncReadBuffer<T>

pub fn into_reply_stream(self) -> impl Stream<Item = Result<Reply>>

Convert a AsyncReadBuffer into Stream<Item = Result<Reply>>.

The implementation of it is use futures::stream::unfold" and every time it will read one reply on this stream (with the help of read_exactandread_line`)

impl<T: AsyncRead + Unpin + 'static> AsyncReadBuffer<T>

pub fn into_command_stream(self) -> impl Stream<Item = Result<Command>>

Convert a AsyncReadBuffer into Stream<Item = Result<Command>>.